adk/vault_watsonx/test/requirements.txt
hvac>=1.1.0
python-dotenv>=0.19.0